Testing

+++ WithoutMultiplexer/Final_Project.c
@@ -263,14 +263,14 @@ void screamDetected(int soundLevel, float waitTime)
 		time1[T1] = 0;
 		while (time1[T1] < waitTime)
 		{}
-		float reading1 = SensorValue(S2);
+		int reading1 = SensorValue(S2);
 		while (time1[T1] < waitTime * 2)
 		{}
-		float reading2 = SensorValue(S2);
+		int reading2 = SensorValue(S2);
 		while (time1[T1] < waitTime * 3)
 		{}
-		float reading3 = SensorValue(S2);
-		average = (reading1 + reading2 + reading3) / 3.0;
+		int reading3 = SensorValue(S2);
+		average = (float)(reading1 + reading2 + reading3) / 3.0;
 	}
 }
